Pause Menu & Ability to skip splash screen
The Pause Menu has not been fully implemented and if there are error/merge conflicts then I will look at it another day. Other than that there are some things that I want to go over like changing the text for the coin counter along with these two similar errors caused by something to do with the portal and thats it. I'll work on stuff like transitions between scenes next time or something like that but cya on Wednesday.