Finished Death System
Im gonna work on a bit more stuff this weekend so we can make a new release that includes all parts of the game. I'll work tmrw to make sure everything works when this is merged in along with adding the rest of the music and I think I know how to make the platform go back up.